backend/requirements.txt
flask 
flask-cors 
pandas 
scikit-learn 
joblib 
geopy
python-dotenv
pymongo
colorlog
flask-restx